+++ DiabloII.Application.Tests/Steps/Items/ItemsApi.cs
@@ -2,7 +2,7 @@
 using System.Threading.Tasks;
 using DiabloII.Application.Requests.Items;
 using DiabloII.Application.Responses.Items;
-using DiabloII.Application.Tests.Startup;
+using DiabloII.Application.Tests.Contexts;
 
 namespace DiabloII.Application.Tests.Steps.Items
 {
