+++ DiabloII.Application.Tests/Steps/Items/Search/SearchItemsSteps.cs
@@ -1,6 +1,6 @@
 ﻿using System.Threading.Tasks;
 using DiabloII.Application.Requests.Items;
-using DiabloII.Application.Tests.Startup;
+using DiabloII.Application.Tests.Contexts;
 using TechTalk.SpecFlow;
 using TechTalk.SpecFlow.Assist;
