+++ DiabloII.Application.Tests/Steps/Suggestions/GetAll/GetAllSuggestionsSteps.cs
@@ -1,6 +1,6 @@
 ﻿using System.Threading.Tasks;
+using DiabloII.Application.Tests.Contexts;
 using DiabloII.Application.Tests.Extensions;
-using DiabloII.Application.Tests.Startup;
 using TechTalk.SpecFlow;
 
 namespace DiabloII.Application.Tests.Steps.Suggestions.GetAll
