+++ DiabloII.Application.Tests/Steps/Suggestions/SuggestionsApi.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Threading.Tasks;
 using DiabloII.Application.Requests.Suggestions;
 using DiabloII.Application.Responses.Suggestions;
-using DiabloII.Application.Tests.Startup;
+using DiabloII.Application.Tests.Contexts;
 
 namespace DiabloII.Application.Tests.Steps.Suggestions
 {
