+++ DiabloII.Application.Tests/Steps/Suggestions/SuggestionsRepository.cs
@@ -2,9 +2,8 @@
 using System.Linq;
 using System.Threading.Tasks;
 using DiabloII.Application.Responses.Suggestions;
-using DiabloII.Application.Tests.Steps.Suggestions;
 
-namespace DiabloII.Application.Tests.Startup
+namespace DiabloII.Application.Tests.Steps.Suggestions
 {
     public class SuggestionsRepository
     {
